+++ src/main/java/team03/mopl/domain/content/batch/common/BatchConfig.java
@@ -5,7 +5,7 @@ import org.springframework.batch.item.ItemWriter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import team03.mopl.domain.content.Content;
-import team03.mopl.domain.content.ContentRepository;
+import team03.mopl.domain.content.repository.ContentRepository;
 
 @Configuration
 @RequiredArgsConstructor
