+++ src/main/java/team03/mopl/domain/content/batch/sports/SportsBatchConfig.java
@@ -13,8 +13,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.web.client.RestTemplate;
 import team03.mopl.domain.content.Content;
-import team03.mopl.domain.content.ContentRepository;
-import team03.mopl.domain.content.batch.common.ApiWriter;
 
 @Configuration
 @RequiredArgsConstructor
