+++ src/main/java/team03/mopl/domain/content/repository/ContentRepository.java
@@ -1,0 +1,13 @@
+package team03.mopl.domain.content.repository;
+
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+import team03.mopl.domain.content.Content;
+import team03.mopl.domain.content.ContentType;
+
+public interface ContentRepository extends JpaRepository<Content, UUID> {
+
+  boolean existsByContentType(ContentType contentType);
+
+  boolean existsByDataId(String dataId);
+}
